ci: Always rebuild kwarc.bib (fixes #27)
Previously, the GitHub Workflow first checked if changes that require a
rebuild of the bibliography file were made. As noted in #27, this
triggered some false negatives when rebuilds that should have happened
did not occur. However it was neccessary with Travis CI, to prevent
cycles of rebuilding the files over and over.

This commit fixes the issue by dropping the check entirely, which is
safe with GitHub Actions when adding "[skip ci]" to the commit message
title as per [1].

[1]
https://docs.github.com/en/actions/managing-workflow-runs/skipping-workflow-runs