generate-html: Handle pubs with "-" in name
Previously, the "voll-ki" project pages were not building correctly. On
closer investigation this turns out to have been an improper splitting
of "pubs" and "type" parameters inside the xslt file. In particular,
they were just concatinated using a "-", and then split again, leading
to projects with a "-" in them being interpreted incorrectly.

This commit updates the behaviour to instead pass them as seperate
parameters. This fixes the "voll-ki" as well as related publication
pages.